requirements.txt
amqp
anyjson
beautifulsoup4
billiard
Django>=2.1
django-jinja
django-widget-tweaks
Jinja2
kombu
Markdown
MarkupSafe
ordered-set
Pygments
python-markdown-math
pytz
requests
shortuuid
sqlparse
vine
djangorestframework
django-tagging
xlsxwriter
django-debug-toolbar
names
django-imagekit
pillow
django-simple-captcha
mkdocs
mkdocs-bootswatch
django-redis
pyyaml
chardet
socketio-client
django-extensions
django-comments-xtd>=2.2
django-notifications-hq>=1.5
django-ipware
psycopg2-binary
uptime
python-dateutil
django-q
pycrypto
djangorestframework_simplejwt
progressbar2